Refactor tests to follow idiomatic Rust inline convention
Move all unit tests from separate src/tests.rs files into #[cfg(test)]
mod tests blocks at the bottom of each source file. This follows the
standard Rust convention of keeping tests alongside the code they cover.

Updated AGENTS.md with testing guidance (key truth #5) and STRUCTURE.md
to reflect the new file layout.
