+++ crates/pokeplanner-core/src/lib.rs
@@ -1,8 +1,6 @@
 pub mod error;
 pub mod job;
 pub mod model;
-#[cfg(test)]
-mod tests;
 
 pub use error::AppError;
 pub use job::{Job, JobId, JobResult, JobStatus};
