+++ crates/pokeplanner-storage/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod json_store;
 pub mod traits;
-#[cfg(test)]
-mod tests;
 
 pub use json_store::JsonFileStorage;
 pub use traits::Storage;
